frontend/requirements.txt
Flask==3.1.1
mcp-use==1.3.3
langchain-ollama==0.3.5
python-dotenv==1.0.0 